Refactored the cache used by ExperienceStore and ImageStore into shared variables.
